drug_agent/requirements.txt
# Drug Discovery Agent Requirements
# ==================================

# Qdrant Cloud Client
qdrant-client>=1.7.0

# HuggingFace Transformers & Embeddings
sentence-transformers>=2.2.0
transformers>=4.35.0
torch>=2.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Progress tracking
tqdm>=4.65.0

# HTTP requests (for HuggingFace downloads)
requests>=2.28.0

# ChEMBL API Client (for drug data fetching)
chembl_webresource_client>=0.10.8

# Development
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: GPU acceleration
# accelerate>=0.24.0

# Optional: HuggingFace Hub (for model downloads)
huggingface-hub>=0.19.0
